Refactor Firebase initialization and update quiz game script references. Replace optimized quiz game script with a new version that includes dynamic imports for improved performance. Enhance button styling for disabled states across various components to ensure consistent user experience. Add error handling for Firebase initialization to improve reliability.
